Fixed tip field getters

+++ public/js/hb-helpers.js
@@ -91,8 +91,8 @@ var helpers = {
   },
 
   dollars: function(pennies) {
-    var cents = pennies == null ? 0 : parseInt(pennies);
-    return utils.isNaN(cents) ? '' : (cents / 100).toFixed(2);
+    var cents = pennies == null ? 0 : parseFloat(pennies); // parse as float incase of partial cents
+    return utils.isNaN(cents) ? '' : (cents / 100).toFixed(2); // partial cents get rounded here
   },
 
   json: function(context) {
